Drop AppLayout.Whatevers and fix vertical margins
