+++ README.md
@@ -2,11 +2,11 @@
 
 A React application to visualize workout data. For more information, see [this blog post](https://danny.codes/blog/workout-analytics).
 
-**URL**: https://thirsty-swanson-f8ea25.netlify.app
+**URL**: https://workout-analytics.netlify.app
 
 ## 🔥 Status
 
-[![Netlify Status](https://api.netlify.com/api/v1/badges/6edc21c6-07e8-4bb5-978b-7674ee04ddd5/deploy-status)](https://app.netlify.com/sites/thirsty-swanson-f8ea25/deploys)
+[![Netlify Status](https://api.netlify.com/api/v1/badges/6edc21c6-07e8-4bb5-978b-7674ee04ddd5/deploy-status)](https://app.netlify.com/sites/workout-analytics/deploys)
 
 ## 🏃‍♂️ Scripts
 
@@ -22,6 +22,6 @@ A React application to visualize workout data. For more information, see [this b
 
 - **Framework**: React (via [create-react-app](https://github.com/facebook/create-react-app))
 - **Hosting**: [Netlify](https://www.netlify.com/)
-- **CI**: [GitHub Actions](https://github.com/xdmorgan/peloton-workouts/actions)
+- **CI**: [GitHub Actions](https://github.com/xdmorgan/workout-analytics/actions)
 - **Testing**: [Jest](https://jestjs.io/) & [react-testing-library](https://testing-library.com/)
 - **Styling**: Sass Modules, [Skeletor](https://www.npmjs.com/package/@skeletor/css), and [Typebeast](https://www.npmjs.com/package/typebeast)
