+++ src/pages/10-totals.jsx
@@ -1,9 +1,10 @@
 import React from "react";
-import { AppLayout } from "../layouts/app-layout";
 import { TRANSFORMED_KEYS } from "../constants";
 import { TotalPiechart } from "../components/total-piechart";
 import { ListWithIcon } from "../components/list-with-icon";
 import { ContentHeader } from "../components/content-header";
+import { ContentSection } from "../components/content-section";
+import { Pagination } from "../components/pagination";
 
 export const meta = {
   route: "/totals",
@@ -27,7 +28,7 @@ export function Page({ allWorkoutData, pageMetadata }) {
       >
         {pageMetadata.title}
       </ContentHeader>
-      <AppLayout.Content divider="after">
+      <ContentSection divider="after">
         <TotalPiechart data={totals.numberOfWorkouts.pieChart}>
           <p className="type-h1-xxl my-2x">{totals.numberOfWorkouts.sum}</p>
           <h2 className="type-h2">Combined Workouts</h2>
@@ -39,8 +40,8 @@ export function Page({ allWorkoutData, pageMetadata }) {
               : " category"}
           </p>
         </TotalPiechart>
-      </AppLayout.Content>
-      <AppLayout.Content divider="after">
+      </ContentSection>
+      <ContentSection divider="after">
         <TotalPiechart data={totals.timeInMinutes.pieChart}>
           <p className="type-h1-xxl my-2x">{totals.timeInMinutes.sum}</p>
           <h2 className="type-h2">Minutes Completed</h2>
@@ -65,8 +66,8 @@ export function Page({ allWorkoutData, pageMetadata }) {
             </ListWithIcon.Item>
           </ListWithIcon>
         </TotalPiechart>
-      </AppLayout.Content>
-      <AppLayout.Content divider="after">
+      </ContentSection>
+      <ContentSection divider="after">
         <TotalPiechart data={totals.distanceInMiles.pieChart}>
           <p className="type-h1-xxl my-2x">{totals.distanceInMiles.sum}</p>
           <h2 className="type-h2">Miles Covered</h2>
@@ -92,8 +93,8 @@ export function Page({ allWorkoutData, pageMetadata }) {
             </ListWithIcon.Item>
           </ListWithIcon>
         </TotalPiechart>
-      </AppLayout.Content>
-      <AppLayout.Content>
+      </ContentSection>
+      <ContentSection>
         <TotalPiechart data={totals.caloriesBurned.pieChart}>
           <p className="type-h1-xxl my-2x">{totals.caloriesBurned.sum}</p>
           <h2 className="type-h2">Calories Burned</h2>
@@ -116,11 +117,10 @@ export function Page({ allWorkoutData, pageMetadata }) {
             </ListWithIcon.Item>
           </ListWithIcon>
         </TotalPiechart>
-      </AppLayout.Content>
-      <AppLayout.Pagination
+      </ContentSection>
+      <Pagination
         previousRoute={pageMetadata.pagination.previous}
         nextRoute={pageMetadata.pagination.next}
-        nextLabel="Activity"
       />
     </>
   );
